+++ validator/client/beacon-api/propose_beacon_block_phase0_test.go
@@ -55,7 +55,7 @@ func TestProposeBeaconBlock_Phase0(t *testing.T) {
 	headers := map[string]string{"Eth-Consensus-Version": "phase0"}
 	jsonRestHandler.EXPECT().Post(
 		gomock.Any(),
-		"/eth/v1/beacon/blocks",
+		"/eth/v2/beacon/blocks",
 		headers,
 		bytes.NewBuffer(marshalledBlock),
 		nil,